+++ CHANGELOG.md
@@ -21,6 +21,21 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+## v0.2.1
+
+Released Tuesday, 26th May 2020.
+
+### New
+
+* Added `addExtension()`
+  - based on the older `addExtensions()` function
+  - name better reflects that we can only add one type of extension at a time
+
+### Deprecated
+
+* `addExtensions()` is now deprecated
+  - use `addExtension()` instead
+
 ## v0.2.0
 
 Released Thursday, 21st May 2020.